pyproject.toml
[project]
name = "assemblyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
]
